Add a dumb substring filter to more quickly focus on commits of interest.
Note that this filtering algorithm matches on the entire displayed text
of each row in the commits table.  It's sensitive to differences in
case, etc, but that's still enough to be useful for quickly finding a
specific commit number or all commits from a specific author.  It also
errs on the side of firing more than once instead of letting any
changes to the input box fall through unhandled, but the filter
executes quickly enough that this doesn't seem to be a problem for our
typical use cases.